tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/046.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: no readystatechange events when adding external scripts </title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		testlib.addScript('', {src:'scripts/include-1.js', onreadystatechange:function() {log( 'readystatechange '+ this.readyState );}}, document.getElementsByTagName('head')[0], false );
		log('end script #1');
	</script>

	<script type="text/javascript">
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'external script #1']);
		t.done();
        }
	onload = t.step_func(test);
	</script>

</body></html>
